termios platforms: make Close() unblock readers
Close() on an *os.File only unblocks readers if the File has a working
SetDeadline implementation. SetDeadline stops working after calling .Fd()
on *os.File(). This commit removes all the calls to .Fd() where it is
feasible, i.e. not in the TakeOver() function. Otherwise, we open the
serial file with syscall.Open, operate on the fd and only later turn it
into an *os.File.